extract `CEditorSound` into separate header file
