+++ src/game/editor/editor.cpp
@@ -32,6 +32,7 @@
 #include <game/localization.h>
 
 #include <game/editor/mapitems/image.h>
+#include <game/editor/mapitems/sound.h>
 
 #include "auto_map.h"
 #include "editor.h"
